Added new class VariantGenotypedVcfFactory. VCF factories refactor
