+++ variation-commons-core/src/main/java/uk/ac/ebi/eva/commons/core/models/factories/VariantVcfFactory.java
@@ -17,13 +17,10 @@
 package uk.ac.ebi.eva.commons.core.models.factories;
 
 import org.apache.commons.lang3.StringUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import uk.ac.ebi.eva.commons.core.models.VariantCoreFields;
 import uk.ac.ebi.eva.commons.core.models.factories.exception.IncompleteInformationException;
 import uk.ac.ebi.eva.commons.core.models.factories.exception.NonVariantException;
-import uk.ac.ebi.eva.commons.core.models.genotype.Genotype;
 import uk.ac.ebi.eva.commons.core.models.pipeline.Variant;
 import uk.ac.ebi.eva.commons.core.models.pipeline.VariantSourceEntry;
 
@@ -34,14 +31,11 @@ import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.TreeMap;
 
 /**
  * Class that parses VCF lines to create Variants.
  */
-public class VariantVcfFactory {
-
-    private static final Logger logger = LoggerFactory.getLogger(VariantVcfFactory.class);
+public abstract class VariantVcfFactory {
 
     /**
      * Creates a list of Variant objects using the fields in a record of a VCF
@@ -177,78 +171,10 @@ public class VariantVcfFactory {
         return secondaryAlternates;
     }
 
-    protected void parseSplitSampleData(Variant variant, String fileId, String studyId, String[] fields,
+    protected abstract void parseSplitSampleData(Variant variant, String fileId, String studyId, String[] fields,
                                         String[] alternateAlleles, String[] secondaryAlternates,
-                                        int alternateAlleleIdx) {
-        String[] formatFields = variant.getSourceEntry(fileId, studyId).getFormat().split(":");
+                                        int alternateAlleleIdx);
 
-        for (int i = 9; i < fields.length; i++) {
-            Map<String, String> map = new TreeMap<>();
-
-            // Fill map of a sample
-            String[] sampleFields = fields[i].split(":");
-
-            // Samples may remove the trailing fields (only GT is mandatory),
-            // so the loop iterates to sampleFields.length, not formatFields.length
-            for (int j = 0; j < sampleFields.length; j++) {
-                String formatField = formatFields[j];
-                String sampleField = processSampleField(alternateAlleleIdx, formatField, sampleFields[j]);
-
-                map.put(formatField, sampleField);
-            }
-
-            // Add sample to the variant entry in the source file
-            variant.getSourceEntry(fileId, studyId).addSampleData(map);
-        }
-    }
-
-    /**
-     * If this is a field other than the genotype (GT), return unmodified. Otherwise,
-     * see {@link VariantVcfFactory#processGenotypeField(int, java.lang.String)}
-     *
-     * @param alternateAlleleIdx current alternate being processed. 0 for first alternate, 1 or more for a secondary alternate.
-     * @param formatField as shown in the FORMAT column. most probably the GT field.
-     * @param sampleField parsed value in a column of a sample, such as a genotype, e.g. "0/0".
-     * @return processed sample field, ready to be stored.
-     */
-    private String processSampleField(int alternateAlleleIdx, String formatField, String sampleField) {
-        if (formatField.equalsIgnoreCase("GT")) {
-            return processGenotypeField(alternateAlleleIdx, sampleField);
-        } else {
-            return sampleField;
-        }
-    }
-
-    /**
-     * Intern the genotype String into the String pool to avoid storing lots of "0/0". In case that the variant is
-     * multiallelic and we are currently processing one of the secondary alternates (T is the only secondary alternate
-     * in a variant like A -> C,T), change the allele codes to represent the current alternate as allele 1. For details
-     * on changing this indexes, see {@link VariantVcfFactory#mapToMultiallelicIndex(int, int)}
-     *
-     * @param alternateAlleleIdx current alternate being processed. 0 for first alternate, 1 or more for a secondary alternate.
-     * @param genotype first field in the samples column, e.g. "0/0"
-     * @return the processed genotype string, as described above (interned and changed if multiallelic).
-     */
-    private String processGenotypeField(int alternateAlleleIdx, String genotype) {
-        boolean isNotTheFirstAlternate = alternateAlleleIdx >= 1;
-        if (isNotTheFirstAlternate) {
-            Genotype parsedGenotype = new Genotype(genotype);
-
-            StringBuilder genotypeStr = new StringBuilder();
-            for (int allele : parsedGenotype.getAllelesIdx()) {
-                if (allele < 0) { // Missing
-                    genotypeStr.append(".");
-                } else {
-                    // Replace numerical indexes when they refer to another alternate allele
-                    genotypeStr.append(String.valueOf(mapToMultiallelicIndex(allele, alternateAlleleIdx)));
-                }
-                genotypeStr.append(parsedGenotype.isPhased() ? "|" : "/");
-            }
-            genotype = genotypeStr.substring(0, genotypeStr.length() - 1);
-        }
-
-        return genotype.intern();
-    }
 
     protected void setOtherFields(Variant variant, String fileId, String studyId, Set<String> ids, float quality, String filter,
                                   String info, String format, int numAllele, String[] alternateAlleles, String line) {
@@ -356,30 +282,11 @@ public class VariantVcfFactory {
 
     protected void checkVariantInformation(Variant variant, String fileId, String studyId)
             throws NonVariantException, IncompleteInformationException {
-        VariantSourceEntry variantSourceEntry = variant.getSourceEntry(fileId, studyId);
-        if (!isVariant(variantSourceEntry)) {
-            throw new NonVariantException("The variant " + variant + " has no alternate allele genotype calls");
+        if (variant.getAlternate().equalsIgnoreCase(variant.getReference())) {
+            throw new NonVariantException("The variant " + variant + " reference and alternate alleles are the same");
+        }
+        if (variant.getAlternate().equals(".")) {
+            throw new NonVariantException("The variant " + variant + " has no alternate allele");
         }
     }
-
-    protected boolean isVariant(VariantSourceEntry variantSourceEntry) {
-        boolean isVariant = false;
-
-        List<Map<String, String>> samplesData = variantSourceEntry.getSamplesData();
-        if (!samplesData.isEmpty()) {
-            if (samplesData.stream().map(m -> m.get("GT")).anyMatch(VariantVcfFactory::genotypeHasAlternateAllele)) {
-                isVariant = true;
-            }
-        }
-
-        return isVariant;
-    }
-
-
-    private static boolean genotypeHasAlternateAllele(String sampleField) {
-        // the alternate allele index could be originally other than 1, but the processGenotypeField method has
-        // updated those indexes so all calls to the alternate allele in this variant has now index 1
-        return Arrays.stream(sampleField.split("[/|]")).anyMatch(allele -> allele.equals("1"));
-    }
-
 }
